refactored folder structure for different components
