+++ src/views/components/splash/CustomButton.tsx
