+++ src/views/components/splash/Footer.tsx
