+++ src/views/components/splash/Header.tsx
