+++ src/views/components/splash/SplashImage.tsx
