+++ src/views/components/splash/SplashText.tsx
