+++ src/views/components/stock/StockLastQuoteDataViz.tsx
@@ -1,5 +1,5 @@
 import { Bar } from 'react-chartjs-2';
-import { lastQuoteInterface } from '../../interfaces/stockInfoInterfaces';
+import { lastQuoteInterface } from '../../../interfaces/stockInfoInterfaces';
 
 type StockLastQuoteDataVizProps = {
   lastQuote: lastQuoteInterface | null
