+++ src/views/containers/AllInfoContainer.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from 'react';
 import { Tabs, Tab } from '@material-ui/core';
 
 import SearchBar from '../components/SearchBar';
-import CompanyInfo from '../components/CompanyInfo';
+import CompanyInfo from '../components/company/CompanyInfo';
 import StockInfoContainer from './StockInfoContainer';
 
 import { stockInfoControllers } from "../../controllers/stockInfoControllers";
